# api-swagger-docs/openapi.yaml
openapi: 3.0.3
info:
  title: ECommerce Application
  description: >-
    This is E-commmerce REST API which provides simple endpoints which can be used by
    anyone to instantly to lauch a backend Ecommerece servers with minimum
    customizations and also helps to scale within less time
    
    - Project Details :
    
        - [Backend Server Code](https://github.com/Vinay-M-G/Simple-Commerece-Backend)
        - [Frontend Application Code](https://github.com/Vinay-M-G/Simple-Commerece--frontend-)
  contact:
    email: vinay.giriswamy@gmail.com
  version: 3.0.0
servers:
  - url: http://{EC2 Server Id}:9001/api/v1
    variables:
      EC2 Server Id:
        default: Add_domain_here
...
paths:
...
  /user/createregistereduser:
    post:
      tags:
        - User Controller
      summary: Allows user to create Account
      description: Allows user to create Account
      requestBody:
        required: true
        description: Request Body for Creating Account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountModel'
      responses:
        '201':
          description: Successfull response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  error: 
                    type: boolean
                    default: false
                  errorMessage:
                    type: string
                    default: null
        '409':
          description: Error Response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  error: 
                    type: boolean
                    default: true
                  errorMessage:
                    type: string
  /user/createguestuser:
    post:
      tags:
        - User Controller
      summary: Allows user to create Account for Guest User
      description: Allows user to create Account for Guest User
      requestBody:
        required: true
        description: Request Body for Creating guest Account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestUserModel'
      responses:
        '201':
          description: Successfull response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  authStatus: 
                    type: string
                    default: Success
                  errorMessage:
                    type: string
                    default: null 
                  guid:
                    type: string
                  error:
                    type: boolean
                    default: false
        '409':
          description: Error response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  authStatus: 
                    type: string
                    default: Failed
                  errorMessage:
                    type: string
                  error:
                    type: boolean
                    default: true            
  /user/reguserlogin:
    post:
      tags:
        - User Controller
      summary: Allows registered user to login
      description: Allows registered user to login
      requestBody:
        required: true
        description: Request Body for Creating guest Account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginModel'
      responses:
        '201':
          description: Successfull response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  authStatus: 
                    type: string
                    default: Success
                  errorMessage:
                    type: string
                    default: null 
                  guid:
                    type: string
                  error:
                    type: boolean
                    default: false
        '409':
          description: Error response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  authStatus: 
                    type: string
                    default: Failed
                  errorMessage:
                    type: string
                  error:
                    type: boolean
                    default: true
  /user/updateaddress:
    post:
      tags:
        - User Controller
      summary: Allows users to update the shipping and billing address
      description: |-
        A single endpoint to update user address for both cart level and user level. Works both for guest and registered user
      parameters:
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Request Body for updating user address
        content:
          application/json:
            schema:
              type: object
              properties:
                line1:
                  type: string
                line2:
                  type: string
                pincode:
                  type: string
                countryCode:
                  type: string
                country: 
                  type: string
                phoneNumber:
                  type: string
                city:
                  type: string
                state:
                  type: string
                addressTitle:
                  type: string
                isBusnisessOrder:
                  type: boolean
                companyName:
                  type: string
                isBillingAddress:
                  type: boolean
      responses:
        '201':
          description: Successfull response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: boolean
                    default: false
        '409':
          description: Error response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: boolean
                    default: true                  
  /user/getaddresslist:
    get:
      tags:
        - User Controller
      summary: provides address values based on the previous address details provided by the user
      description: provides address details based on previous history
      parameters:
        - name: busnisessType
          in: query
          description: |-
            - adrress recommendations will be provided based on B2B or B2C
          required: true
          schema:
            type: string
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string    
      responses:
        '200':
          description: Successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressModel'
        '409':
          description: Error response
          content: 
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: boolean
                    default: true                  
...
components:
  schemas:
...
    GuestUserModel:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
...
    CreateAccountModel:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    UserLoginModel:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    AddressModel:
      type: object
      properties: 
        addressValues:
          type: array
          items:
            properties:
              pk:
                type: string
              emailId:
                type: string
              line1:
                type: string
              line2:
                type: string
              pincode:
                type: string
              countryCode:
                type: string
              country: 
                type: string
              phoneNumber:
                type: string
              city:
                type: string
              state:
                type: string
              addressType:
                type: string
                enum:
                  - BILLING
                  - SHIPPING
              addressTitle:
                type: string
              busnisessType:
                type: string
                enum:
                  - B2C
                  - B2B
              companyName:
                type: string
              updatedAt:
                type: string
        error:
          type: boolean
        emailId:
          type: string
        message:
          type: string
...